pytesseract-svc/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
pytesseract==0.3.10
Pillow==9.5.0
Werkzeug==2.0.3
